Remove Pirate Doom 2 support
Looking closer at Pirate Doom 2, its readme says its contents are not to be used as a base for modification. I've left the map checksum in for others who might want to add a status bar for it for private use.

I'm not adding support for anything else until I review all I've done so far and removed anything else that shouldn't be here. I should know better than to not check these things.